server/requirements.txt
# Server requirements
wit==5.1.0
python-dotenv==0.13.0
python-multipart==0.0.5
